requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
influxdb3-python==0.6.1
numpy==2.0.0
pandas==2.2.2
pyarrow==16.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
reactivex==4.0.4
requests==2.32.3
setuptools==70.2.0
six==1.16.0
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
